@@ -9,7 +9,7 @@
     <AppDesignerFolder>Properties</AppDesignerFolder>
     <RootNamespace>UncomplicatedCustomBots</RootNamespace>
     <AssemblyName>UncomplicatedCustomBots</AssemblyName>
-    <LangVersion>9.0</LangVersion>
+    <LangVersion>12</LangVersion>
     <TargetFrameworkVersion>v4.8</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <Deterministic>true</Deterministic>
