@@ -1,12 +1,21 @@
-﻿using Exiled.API.Features;
+﻿using LabApi.Events.Arguments.ServerEvents;
+using LabApi.Features.Wrappers;
+using LabApi.Loader.Features.Paths;
+using MEC;
+using NetworkManagerUtils.Dummies;
 using System;
 using System.Collections.Generic;
+using System.IO;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+using UncomplicatedCustomBots.API.Extensions;
 using UncomplicatedCustomBots.API.Features;
+using UncomplicatedCustomBots.API.Features.Components;
+using UncomplicatedCustomBots.API.Managers;
 using UnityEngine;
-using EventTarget = Exiled.Events.Handlers.Server;
+using UnityEngine.AI;
+using UnityEngine.UIElements;
+using EventTarget = LabApi.Events.Handlers.ServerEvents;
+using Logger = LabApi.Features.Console.Logger;
 
 namespace UncomplicatedCustomBots.Events.Internal
 {
@@ -14,32 +23,42 @@ namespace UncomplicatedCustomBots.Events.Internal
     {
         public static void Register()
         {
-            EventTarget.WaitingForPlayers += AddRoomNodesOnWaitingForPlayers;
+            EventTarget.MapGenerated += OnMapGenerated;
+            EventTarget.RoundStarting += OnRoundStarted;
         }
 
         public static void Unregister()
         {
-            EventTarget.WaitingForPlayers -= AddRoomNodesOnWaitingForPlayers;
+            EventTarget.MapGenerated -= OnMapGenerated;
+            EventTarget.RoundStarting -= OnRoundStarted;
         }
 
-        public static void AddRoomNodesOnWaitingForPlayers()
+        public static void OnMapGenerated(MapGeneratedEventArgs ev)
         {
-            foreach (var room in Room.List)
+            foreach (Room room in Room.List)
             {
-                var ways = new List<Room>();
+                List<Room> ways = [];
 
-                //thats shit i know
-                foreach (var way in Room.List)
+                foreach (Room way in Room.List)
                 {
                     if (Vector3.Distance(room.Position, way.Position) > 15)
-                    {
                         continue;
-                    }
 
                     ways.Add(way);
                 }
 
                 RoomNode.List.Add(new RoomNode(room, ways.ToArray()));
+            }
+        }
+
+        public static void OnRoundStarted(RoundStartingEventArgs ev)
+        {
+            if (Player.ReadyList.Count() >= Plugin.Instance.Config.MaxPlayers)
+                return;
+
+            for (int i = 0; i < Plugin.Instance.Config.MaxBots; i++)
+            {
+                new Bot();
             }
         }
     }
