@@ -1,24 +1,19 @@
-﻿using Mirror;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using UnityEngine;
 
 namespace UncomplicatedCustomBots.API.Features.Components
 {
-    public class BotComponent : NetworkBehaviour
+    public class BotComponent : MonoBehaviour
     {
         private Bot _bot;
 
         public void Initialize(Bot bot)
         {
-            _bot ??= bot;
+            _bot = bot;
         }
 
         public void Update()
         {
-            _bot?.State.Update();
+            _bot?.Update();
         }
     }
 }
