@@ -1,11 +1,7 @@
-﻿using Exiled.API.Enums;
-using Exiled.API.Features;
-using System;
+﻿using LabApi.Features.Wrappers;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using UncomplicatedCustomBots.API.Features;
+using UncomplicatedCustomBots.API.Extensions;
 using UncomplicatedCustomBots.API.Interfaces;
 using UnityEngine;
 
@@ -13,7 +9,7 @@ namespace UncomplicatedCustomBots.API
 {
     public readonly struct RoomTarget : ITarget
     {
-        public RoomTarget(Vector3 offset, RoomType room)
+        public RoomTarget(Vector3 offset, Room room)
         {
             Offset = offset;
             Room = room;
@@ -21,8 +17,10 @@ namespace UncomplicatedCustomBots.API
 
         public Vector3 Offset { get; }
 
-        public RoomType Room { get; }
+        public Room Room { get; }
 
-        public Queue<Room> GetWay(Room room) => Exiled.API.Features.Room.Get(Entity.Position).GetRoomNode().GetWay(room);
+        public Queue<Room> GetWay(Room room) => Room.GetRoomAtPosition(Room.Position).GetRoomNode().GetWay(room);
+        public static List<Room> GetAdjacentRooms(Room room) => room.AdjacentRooms.ToList();
+        public static List<Room> GetPath(Room roomStart, Room roomEnd) => Room.FindPath(roomStart, roomEnd); 
     }
 }
